The first working version
Currently there are 3 AI ships in different games being printed on the screen at the same time. Once the spacebar is pressed, the drawing function stops updating frames and 200 generations are advanced forward into the neural network (the progress can be seen in the console by pressing ctrl + shift + j).